pyproject.toml
[tool.poetry]
name = "fletpython"
version = "0.1.0"
description = ""
authors = ["kutya001 <kutmanomuraliev012@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flet = "^0.9.0"
psycopg2-binary = "^2.9.7"
environ = "^1.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
